refactor(mix): use file path
